Add `json_loads()`, `json_load()` and `json_save()` to common, create global config.schema object
